requirements.txt
gradio
easyocr
opencv-python-headless
pillow
pandas
jiwer
pytz
python-dotenv
langchain-ollama
